Add basic framework to extract YAML witnesses
